added new component TranslatinTextArea
added new component TranslationTextArea and added in it repeated logic from Translation and ListenWrite components
