+++ client/src/components/ListenWrite/index.tsx
@@ -5,8 +5,7 @@ import { RootState } from '../../reducers';
 import { ISentence } from '../../types/Sentence';
 import PracticeBtns from '../PracticeBtns';
 import Audio from '../Audio/Audio';
-import { setCurrentTranslation } from '../../actions/practice/practiceSentencesActions';
-import { useAppDispatch } from '../../store';
+import TranslationTextArea from '../TranslationTextArea';
 
 const ListenWrite: ListenWriteProps = ({
   onSubmit,
@@ -17,31 +16,6 @@ const ListenWrite: ListenWriteProps = ({
   const [disablePlay, setDisablePlay] = useState(false);
   const [playAudio, setPlayAudio] = useState(false);
   const [audioSrc, setAudioSrc] = useState('./audio1.mp3');
-
-  const dispatch = useAppDispatch();
-  const [translationInput, setTranslationInput] = useState('');
-  const [focusedClass, setFocusedClass] = useState('');
-
-  useEffect(() => {
-    setTranslationInput('');
-  }, [currentSentence]);
-
-  const onFormChange = (e: any) => {
-    if (e.key === 'Enter') {
-      e.preventDefault();
-      onSubmit();
-    }
-  }
-
-  const handleOnContinue = () => {
-    onContinue();
-    setTranslationInput('');
-  }
-
-  const handleTranslationChange = (e: any) => {
-    setTranslationInput(e);
-    dispatch(setCurrentTranslation(e));
-  }
 
   useEffect(() => {
     if (currentSentence.audio) {
@@ -69,17 +43,9 @@ const ListenWrite: ListenWriteProps = ({
     <div>
       <button disabled={disablePlay} onClick={handlePlayAudio}>Play</button>
       <Audio play={playAudio} src={audioSrc} />
-      <textarea
-        className={`txt-area ${focusedClass}`}
-        value={translationInput}
-        onChange={(e) => handleTranslationChange(e.target.value)}
-        onKeyDown={onFormChange}
-        onFocus={() => setFocusedClass('focused')}
-        onBlur={() => setFocusedClass('')}
-        placeholder="Type in German"
-      />
+      <TranslationTextArea onSubmit={onSubmit} />
       <div>
-        <PracticeBtns showContinue={showContinue} onContinue={handleOnContinue} onSubmit={onSubmit} />
+        <PracticeBtns showContinue={showContinue} onContinue={onContinue} onSubmit={onSubmit} />
       </div>
     </div>
   )
