+++ client/src/components/TranslationTextArea/types.ts
@@ -1,0 +1,5 @@
+import type { FC } from 'react';
+
+export type ITranslationTextArea = FC<{
+  onSubmit: () => void;
+}>;
